+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/JavadocStyleCheckTest.java
@@ -130,7 +130,7 @@ public class JavadocStyleCheckTest
     }
 
     @Test
-    public void testFirstSentenceFormat() throws Exception {
+    public void testJavadocStyleFirstSentenceFormatOne() throws Exception {
         final String[] expected = {
             "23: " + getCheckMessage(MSG_NO_PERIOD),
             "33: " + getCheckMessage(MSG_NO_PERIOD),
@@ -139,18 +139,44 @@ public class JavadocStyleCheckTest
             "65: " + getCheckMessage(MSG_NO_PERIOD),
             "70: " + getCheckMessage(MSG_NO_PERIOD),
             "75: " + getCheckMessage(MSG_NO_PERIOD),
-            "166: " + getCheckMessage(MSG_NO_PERIOD),
-            "199: " + getCheckMessage(MSG_NO_PERIOD),
-            "325: " + getCheckMessage(MSG_NO_PERIOD),
-            "335: " + getCheckMessage(MSG_NO_PERIOD),
-            "339: " + getCheckMessage(MSG_NO_PERIOD),
-            "346: " + getCheckMessage(MSG_NO_PERIOD),
-            "358: " + getCheckMessage(MSG_NO_PERIOD),
-            "371: " + getCheckMessage(MSG_NO_PERIOD),
         };
 
         verifyWithInlineConfigParser(
-                getPath("InputJavadocStyle3.java"), expected);
+                getPath("InputJavadocStyleFirstSentenceFormatOne.java"), expected);
+    }
+
+    @Test
+    public void testJavadocStyleFirstSentenceFormatTwo() throws Exception {
+        final String[] expected = {
+            "66: " + getCheckMessage(MSG_NO_PERIOD),
+            "99: " + getCheckMessage(MSG_NO_PERIOD),
+        };
+
+        verifyWithInlineConfigParser(
+            getPath("InputJavadocStyleFirstSentenceFormatTwo.java"), expected);
+    }
+
+    @Test
+    public void testJavadocStyleFirstSentenceFormatThree() throws Exception {
+        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
+
+        verifyWithInlineConfigParser(
+            getPath("InputJavadocStyleFirstSentenceFormatThree.java"), expected);
+    }
+
+    @Test
+    public void testJavadocStyleFirstSentenceFormatFour() throws Exception {
+        final String[] expected = {
+            "40: " + getCheckMessage(MSG_NO_PERIOD),
+            "50: " + getCheckMessage(MSG_NO_PERIOD),
+            "54: " + getCheckMessage(MSG_NO_PERIOD),
+            "61: " + getCheckMessage(MSG_NO_PERIOD),
+            "73: " + getCheckMessage(MSG_NO_PERIOD),
+            "86: " + getCheckMessage(MSG_NO_PERIOD),
+        };
+
+        verifyWithInlineConfigParser(
+            getPath("InputJavadocStyleFirstSentenceFormatFour.java"), expected);
     }
 
     @Test
